+++ src/core/service-worker.js
@@ -8,6 +8,7 @@
 import { MessageHandler, MESSAGE_TYPES } from './message-handler.js'
 import { PinboardService } from '../features/pinboard/pinboard-service.js'
 import { LocalBookmarkService } from '../features/storage/local-bookmark-service.js'
+import { SyncBookmarkService } from '../features/storage/sync-bookmark-service.js'
 import { FileBookmarkService } from '../features/storage/file-bookmark-service.js'
 import { InMemoryFileBookmarkAdapter } from '../features/storage/file-bookmark-storage-adapter.js'
 import { MessageFileBookmarkAdapter, ensureOffscreenDocument } from '../features/storage/message-file-bookmark-adapter.js'
@@ -152,6 +153,7 @@ class HoverboardServiceWorker {
     const tagService = this.messageHandler.tagService
     const pinboardProvider = new PinboardService(tagService)
     const localProvider = new LocalBookmarkService(tagService)
+    const syncProvider = new SyncBookmarkService(tagService)
 
     let fileAdapter = new InMemoryFileBookmarkAdapter()
     const storage = await chrome.storage.local.get(['hoverboard_file_storage_configured', 'hoverboard_file_storage_path'])
@@ -179,6 +181,7 @@ class HoverboardServiceWorker {
       pinboardProvider,
       localProvider,
       fileProvider,
+      syncProvider,
       storageIndex,
       getDefaultStorageMode
     )